Assign a sequential ID variable to a dataframe

This function assigns a sequential ID number to distinct rows in a dataframe based on the variables selected in vars.

Usage

assign_id(
  df,
  id_name,
  prefix = NULL,
  vars = colnames(df),
  seq_start = 1,
  digits = NULL,
  promote = TRUE
)

Arguments

df

A dataframe.

id_name

A name for the new ID variable.

prefix

An optional prefix of characters to add to all ID values.

vars

The variable names whose unique values the new ID will be based on. Defaults to colnames(df).

seq_start

An integer for the beginning of the ID sequence. Defaults to 1.

digits

An integer for the number of digits in the ID (including leading zeros). Defaults to nchar(nrow(df) + seq_start).

promote

A logical value. If TRUE, move the ID column to the first position in df.

Value

A dataframe.

Examples

df <- data.frame(
  row_id = paste("r", 1:24),
  a = rep(c("a", "b", "c"), times = 8),
  b = c(1, 1, 2, 2),
  c = c(TRUE, FALSE)
)

df
#>    row_id a b     c
#> 1     r 1 a 1  TRUE
#> 2     r 2 b 1 FALSE
#> 3     r 3 c 2  TRUE
#> 4     r 4 a 2 FALSE
#> 5     r 5 b 1  TRUE
#> 6     r 6 c 1 FALSE
#> 7     r 7 a 2  TRUE
#> 8     r 8 b 2 FALSE
#> 9     r 9 c 1  TRUE
#> 10   r 10 a 1 FALSE
#> 11   r 11 b 2  TRUE
#> 12   r 12 c 2 FALSE
#> 13   r 13 a 1  TRUE
#> 14   r 14 b 1 FALSE
#> 15   r 15 c 2  TRUE
#> 16   r 16 a 2 FALSE
#> 17   r 17 b 1  TRUE
#> 18   r 18 c 1 FALSE
#> 19   r 19 a 2  TRUE
#> 20   r 20 b 2 FALSE
#> 21   r 21 c 1  TRUE
#> 22   r 22 a 1 FALSE
#> 23   r 23 b 2  TRUE
#> 24   r 24 c 2 FALSE

assign_id(
  df,
  id_name = "id",
  prefix = "id",
  vars = c("a", "b", "c")
)
#>      id row_id a b     c
#> 3  id01    r 1 a 1  TRUE
#> 9  id02    r 2 b 1 FALSE
#> 23 id03    r 3 c 2  TRUE
#> 5  id04    r 4 a 2 FALSE
#> 11 id05    r 5 b 1  TRUE
#> 17 id06    r 6 c 1 FALSE
#> 7  id07    r 7 a 2  TRUE
#> 13 id08    r 8 b 2 FALSE
#> 19 id09    r 9 c 1  TRUE
#> 1  id10   r 10 a 1 FALSE
#> 15 id11   r 11 b 2  TRUE
#> 21 id12   r 12 c 2 FALSE
#> 4  id01   r 13 a 1  TRUE
#> 10 id02   r 14 b 1 FALSE
#> 24 id03   r 15 c 2  TRUE
#> 6  id04   r 16 a 2 FALSE
#> 12 id05   r 17 b 1  TRUE
#> 18 id06   r 18 c 1 FALSE
#> 8  id07   r 19 a 2  TRUE
#> 14 id08   r 20 b 2 FALSE
#> 20 id09   r 21 c 1  TRUE
#> 2  id10   r 22 a 1 FALSE
#> 16 id11   r 23 b 2  TRUE
#> 22 id12   r 24 c 2 FALSE