Check date order

Check if dates are ordered from earliest to latest.

Usage

check_date_order(...)

Arguments

...

Date vectors entered in their expected order, earliest to latest.

Value

A logical vector the same length as the inputs.

Details

Any number of date vectors can be checked. Elements at the same index are compared. NAs are ignored.

Examples

dates <- seq.Date(
  as.Date("2023-01-01", "%Y-%m-%d"),
  as.Date("2023-12-31", "%Y-%m-%d"),
  by = "day"
)
dates <- c(dates, rep(NA, times = 10))

n_rows <- 500

df <- data.frame(
  d1 = sample(dates, n_rows, replace = TRUE),
  d2 = sample(dates, n_rows, replace = TRUE),
  d3 = sample(dates, n_rows, replace = TRUE),
  d4 = sample(dates, n_rows, replace = TRUE)
)

df$ordered <- check_date_order(df$d1, df$d2, df$d3, df$d4)